spring security - JWT token

+++ payment-backend/payment-backend/src/main/java/com/example/entity/User.java
@@ -12,59 +12,11 @@ public class User {
   
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Id
-  private int id;
+  private Long id;
   private String name;
   private String email;
   private String mobile;
   private String password;
-
-  // public User(){
-    
-  // }
-
-  // public User(int id, String userName, String email, String mobile, String password) {
-  //   this.id = id;
-  //   this.userName = userName;
-  //   this.email = email;
-  //   this.mobile = mobile;
-  //   this.password = password;
-  // }
-  // public int getId() {
-  //   return id;
-  // }
-  // public void setId(int id) {
-  //   this.id = id;
-  // }
-  // public String getUserName() {
-  //   return userName;
-  // }
-  // public void setUserName(String userName) {
-  //   this.userName = userName;
-  // }
-  // public String getEmail() {
-  //   return email;
-  // }
-  // public void setEmail(String email) {
-  //   this.email = email;
-  // }
-  // public String getMobile() {
-  //   return mobile;
-  // }
-  // public void setMobile(String mobile) {
-  //   this.mobile = mobile;
-  // }
-  // public String getPassword() {
-  //   return password;
-  // }
-  // public void setPassword(String password) {
-  //   this.password = password;
-  // }
-
-  // @Override
-  // public String toString() {
-  //   return "User [id=" + id + ", userName=" + userName + ", email=" + email + ", mobile=" + mobile + ", password="
-  //       + password + "]";
-  // }
 
   
 }
